fix: 유저 정보 저장 관련 로직 변경 #254
- 유저 정보를 저장하는 로직을 header에서 UserAuthSection으로 변경하였습니다.
