+++ frontend/src/components/Header/components/HeaderProfileButton.tsx
@@ -1,7 +1,3 @@
-import { useEffect } from "react";
-import { useNavigate } from "react-router-dom";
-import axios from "axios";
-import { useToast } from "@/components/Toast/useToast";
 import HeaderProfileModal from "./HeaderProfileModal";
 import ProfileSmall from "@/assets/imgs/profileSmall.png";
 
@@ -11,39 +7,6 @@ interface HeaderProfileButtonProps {
 }
 
 const HeaderProfileButton = ({ isProfileClicked, setIsProfileClicked }: HeaderProfileButtonProps) => {
-  const navigate = useNavigate();
-  const showToast = useToast();
-
-  useEffect(() => {
-    const accessToken = localStorage.getItem("token");
-    if (!accessToken) {
-      showToast({ message: "로그인이 필요해요.", type: "alert" });
-      navigate("/userauth");
-    } else {
-      const username = localStorage.getItem("username");
-      const email = localStorage.getItem("email");
-      if (!username || !email) {
-        axios
-          .get(`${import.meta.env.VITE_API_SERVER_URL}/profile`, {
-            headers: {
-              Authorization: accessToken
-            }
-          })
-          .then((result) => {
-            console.log(result);
-            localStorage.setItem("username", result.data.username);
-            localStorage.setItem("email", result.data.email);
-          })
-          .catch((error) => {
-            if (error.response.status === 401) {
-              showToast({ message: "로그인이 만료되었어요.", type: "alert" });
-              navigate("/userauth");
-            }
-          });
-      }
-    }
-  }, []);
-
   return (
     <>
       <button type="button" className="" onClick={() => setIsProfileClicked(!isProfileClicked)}>
