+++ frontend/src/hooks/useAuth.ts
@@ -1,0 +1,40 @@
+import { useEffect } from "react";
+import { useNavigate } from "react-router-dom";
+import axios from "axios";
+import { useToast } from "@/components/Toast/useToast";
+
+const useAuth = () => {
+  const navigate = useNavigate();
+  const showToast = useToast();
+
+  useEffect(() => {
+    const accessToken = localStorage.getItem("token");
+    if (!accessToken) {
+      showToast({ message: "로그인이 필요해요.", type: "alert" });
+      navigate("/userauth");
+    } else {
+      const username = localStorage.getItem("username");
+      const email = localStorage.getItem("email");
+      if (!username || !email) {
+        axios
+          .get(`${import.meta.env.VITE_API_SERVER_URL}/profile`, {
+            headers: {
+              Authorization: accessToken
+            }
+          })
+          .then((response) => {
+            localStorage.setItem("username", response.data.username);
+            localStorage.setItem("email", response.data.email);
+          })
+          .catch((error) => {
+            if (error.response?.status === 401) {
+              showToast({ message: "로그인이 만료되었어요.", type: "alert" });
+              navigate("/userauth");
+            } else showToast({ message: "프로필을 불러오는데 문제가 발생했어요", type: "alert" });
+          });
+      }
+    }
+  }, [navigate, showToast]);
+};
+
+export default useAuth;
